% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_dataManipulation.R
\name{fn_colNamesToDataDict}
\alias{fn_colNamesToDataDict}
\title{Change column names to data dictionary names}
\usage{
fn_colNamesToDataDict(dt_, tbl_dataDict)
}
\description{
Change column names to data dictionary names
}
