% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_splines.R
\name{fn_cubicSplines_ns}
\alias{fn_cubicSplines_ns}
\title{Create a basis for a natural cubic spline}
\usage{
fn_cubicSplines_ns(x, knots, varName, Boundary.knots = NULL)
}
\arguments{
\item{x}{the vector of data}

\item{knots}{the inner knots}

\item{varName}{the prefix to be used for the column names}

\item{Boundary.knots}{the boundary knots.  Can be NULL in which cases the
0.001 and 0.999 percentiles are used.}
}
\value{
a list: bx_ the basis as a sparse matrix; cnames the column names
}
\description{
Create a basis for a natural cubic spline.  Based on the ns function which uses
the qr decomposition which gets very slow when there is lots of data.
}
