% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_splines.R
\name{fn_cubicSplines_wg}
\alias{fn_cubicSplines_wg}
\title{create basis for a cubic spline manually (Wood pp126 & 127, Wahba and Gu)}
\usage{
fn_cubicSplines_wg(x, knots)
}
\description{
create basis for a cubic spline manually (Wood pp126 & 127, Wahba and Gu)
}
