% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_performance.R
\name{fn_deviance}
\alias{fn_deviance}
\title{Calculate proportion of deviance explained}
\usage{
fn_deviance(y, mu, wt = NULL, family = NULL)
}
\arguments{
\item{y}{numeric vector of actual values}

\item{mu}{A vector of predictions}

\item{wt}{weights (for example, exposure)}

\item{family}{family (one of; gaussian, binomial, poisson, gamma)}
}
\value{
deviance, null deviance, proportion explained
}
\description{
The functions for calculating the deviance are taken from the glm.fit function and
the relevant c code.  Results have not been tested against glm output.  For poisson family
y and mu vectors need to be claims counts and not frequencies
}
