% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_glam.R
\name{fn_earthReduceSize}
\alias{fn_earthReduceSize}
\title{Reduce size of earth object}
\usage{
fn_earthReduceSize(x)
}
\arguments{
\item{x}{an earth object}
}
\value{
a reduced size earth object
}
\description{
Reduce size of earth object by dropping the slots we do not need, especially
the bx (we will create our own on the full data) and any predictions etc
}
