% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_dataManipulation.R
\name{fn_explainsTooMuch}
\alias{fn_explainsTooMuch}
\title{Check which features explain too much}
\usage{
fn_explainsTooMuch(dt_, varsToCheck = NULL, targetVar = NULL)
}
\arguments{
\item{dt_}{A data table.}

\item{varsToCheck}{A vector of features for this function to check.}

\item{targetVar}{The target variable (currently binomial 0-1 response only).}
}
\value{
A data table with two columns; varName, the feature names and n_problemObs,
  the number of obervations which perfectly pedict the target variable.
}
\description{
Returns a dataframe of variables which explain too much.
}
\details{
For a binomial response this function goes through each feature and checks if
any values of the feature perfectly classifies 25 or more examples.  In practice
we have found that when this is the case, we are some times dealing with feature
which is actually a proxy for the dependent variable or should not be used in the
analysis for some other reason (such as operational).  Such variables can be an
example of what is called ``leakage''.
}
