% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_glam.R
\name{fn_extractModelData}
\alias{fn_extractModelData}
\title{Extract data for modelling or scoring}
\usage{
fn_extractModelData(fname_data, fname_vars, model_ref,
  vars_notToUseAdditional = NULL, folds, n_max, whereStatement)
}
\arguments{
\item{fname_data}{filename of the RData file where dt_all is stored}

\item{fname_vars}{filename of the RData file where the various vars_ 
variables are stored; i.e. vars_uniqueID, vars_weightAll, vars_fold,
vars_neverToUse, vars_indNotToUse, vars_targetAll, vars_missingInds}

\item{vars_notToUseAdditional}{any additional variables not to use besides
those defined in: vars_neverToUse, vars_indNotToUse}

\item{folds}{folds to extract.  If 'all' then all}

\item{n_max}{(e.g. 1:7, 'all') maximum number of examples.
Bites if n_max is less than the number of examples in folds.
Expected use is to limit size of training data when doing mars.
If want to use all folds set n_max = 'all'}

\item{whereStatement}{a statement that returns TRUE / FALSE for 
each example, using any columns in dt_all.  For example:
lossLimit >= 225000}
}
\value{
dt_
}
\description{
Extracts data needed for modelling, sampling the rows and limiting columns
as required.  Saves details of variables used to paste0(model_ref ,'_varsUsed.RData')
}
