% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_hccv.R
\name{fn_hc}
\alias{fn_hc}
\title{Works on one hccv to return a data table of scores}
\usage{
fn_hc(dt_all, highCardVar, targetVar, exposureVar, foldVar, foldsWithTarget)
}
\arguments{
\item{dt_all}{the name of the dataset}

\item{highCardVar}{a vector of the hccv's}

\item{targetVar}{the target variable}

\item{exposureVar}{the exposure variable}

\item{foldVar}{the fold variable}

\item{foldsWithTarget}{folds which have the target}
}
\value{

}
\description{
Works on one hccv to return a data table of scores
}
