% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_splines.R
\name{fn_hinge}
\alias{fn_hinge}
\title{Create one linear spline basis function}
\usage{
fn_hinge(x, knot, lessOrMore)
}
\arguments{
\item{x}{a vector of data from which to create the hinge}

\item{knot}{the value at which to make the hinge / knot}

\item{lessOrMore}{'less' or 'more'.  Determines if hinge is positive above (more)
or below ('less') the knot}
}
\value{
the value of the hinge at each point of the original vector of data
}
\description{
Function to a create hinge.  If the knot is passed in with a negative sign
a hinge from that point and below is created.  Otherwise from that point
and above
}