% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_splines.R
\name{fn_linearSplines}
\alias{fn_linearSplines}
\title{A function to create linear spline basis functions at given knots}
\usage{
fn_linearSplines(underlying, varName, knots, lessOrMore)
}
\arguments{
\item{underlying}{an underlying vector of data}

\item{varName}{the variable name of the data in the vector}

\item{knots}{the points at which we require the hinges to be}

\item{lessOrMore}{'less' means the hinge will go upwards for values less than
the knot and be zero above.  'more' is vice verca}
}
\value{
a data frame with as many columns as needed for the basis
}
\description{
A function to create linear spline basis functions at given knots
}

