% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_hccv.R
\name{fn_logit}
\alias{fn_logit}
\title{Logit function}
\usage{
fn_logit(x)
}
\arguments{
\item{x}{The mean prediction for which the offset (using the logit) is required}
}
\description{
Logit function to create offset when the mean prediction from a model is known
}
