% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_dataManipulation.R
\name{fn_missingValuesImpute}
\alias{fn_missingValuesImpute}
\title{Impute missing values}
\usage{
fn_missingValuesImpute(dt_, meanValues)
}
\arguments{
\item{dt_}{a data.table over which to impute missing values for
some given numeric or logical variables}

\item{meanValues}{a named vector of variables and mean non-missing
values}
}
\value{
a data.table with missing values imputed - modified inplace
}
\description{
Impute missing values
}

