% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_glam.R
\name{fn_modelMatrix}
\alias{fn_modelMatrix}
\title{Create model matrix}
\usage{
fn_modelMatrix(model_ref, dt_, bln_create_checkncols, bln_test_ncols)
}
\arguments{
\item{model_ref}{the model reference}

\item{dt_}{the data table from which to create the model matrix}

\item{bln_create_checkncols}{(TRUE or FALSE) save the number of columns
in the model matrix from numerics and from categoricals}

\item{bln_test_ncols}{(TRUE or FALSE) test the number of columns
in the model matrix based on previously expected number}
}
\value{
a reduced size earth object
}
\description{
Create model matrix based on saved earth object and categorical
variables chosen by IV
}
