% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_performance.R
\name{fn_normalizedGini}
\alias{fn_normalizedGini}
\title{Calculate normalised gini}
\usage{
fn_normalizedGini(actual, pred)
}
\arguments{
\item{actual}{vector of actual responses}

\item{pred}{vector of predicted responses}
}
\value{
normalised gini
}
\description{
Calculate normalised gini
}
