% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_dataManipulation.R
\name{fn_percentMissing}
\alias{fn_percentMissing}
\title{Check percent missing}
\usage{
fn_percentMissing(dt_)
}
\arguments{
\item{dt_}{a data table with columns to be checked for missings}
}
\value{
a data table with the following rows: varName, class,
    nMissing and pctMissing.  Only outputs rows for variables with
    any missing and outputs in order of descending number missing.
}
\description{
Check percent missing.  The percent missing is round to 0.01%.
}
