% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_admin.R
\name{fn_pkgInstallLoad}
\alias{fn_pkgInstallLoad}
\title{Install and load packages}
\usage{
fn_pkgInstallLoad(x)
}
\arguments{
\item{x}{a vector of packages}
}
\description{
Check if packages are installed and if not then installs them.  See
# http://stackoverflow.com/questions/9341635/check-for-installed-packages-before-running-install-packages
}
