% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_glam.R
\name{fn_prepareTrainingData}
\alias{fn_prepareTrainingData}
\title{Prepare training data}
\usage{
fn_prepareTrainingData(fname_data, fname_vars, model_ref,
  vars_notToUseAdditional = NULL, train_folds, n_train_max,
  whereStatement_train)
}
\arguments{
\item{fname_data}{filename of the RData file where dt_all is stored}

\item{fname_vars}{filename of the RData file where the various vars_ 
variables are stored; i.e. vars_uniqueID, vars_weightAll, vars_fold,
vars_neverToUse, vars_indNotToUse, vars_targetAll, vars_missingInds}

\item{vars_notToUseAdditional}{any additional variables not to use besides
those defined in: vars_neverToUse, vars_indNotToUse}

\item{train_folds}{training folds - typically 1:7}

\item{n_train_max}{maximum number of training examples.  Bites if 
n_train_max is less than the number of examples in train_folds}

\item{whereStatement_train}{a statement that returns TRUE / FALSE for 
each example, using any columns in dt_all.  For example:
lossLimit >= 225000}
}
\value{
dt_train
}
\description{
Prepares the training data, sampling the rows and limiting columns
as required.  Saves details of variables used to paste0(model_ref ,'_varsUsed.RData')
}
