% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_performanceCharts.R
\name{fn_qtileAvE}
\alias{fn_qtileAvE}
\title{Actual versus expected by quantile of prediction}
\usage{
fn_qtileAvE(dt_toUse, train_folds, val_folds, actual, predicted, ex = NULL,
  qtiles = NULL, bln_abline = FALSE, dirROutput)
}
\arguments{
\item{dt_toUse}{A data table containing the variables below and "fold"}

\item{train_folds}{vector of the training folds}

\item{val_folds}{vector of the validation folds}

\item{actual}{The variable name for Actual}

\item{predicted}{The variable name for the prediction or score}

\item{ex}{Exposure measure to divide the sum of predictions at each qtile.  If NULL
this is assumed to be 1 for each example}

\item{qtiles}{If not NULL, these are used for the breaks}

\item{bln_abline}{If TRUE, a line x = y will be added}

\item{dirROutput}{the output directory}
}
\value{
Nothing returned to R, but graph output to dirROutput
}
\description{
Actual versus expected by quantile of prediction
}

