% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_glam.R
\name{fn_score}
\alias{fn_score}
\title{Scoring / predictions}
\usage{
fn_score(model_ref, model_prefix, fname_data, folds_train, whereStatement,
  rebase = FALSE, verbose = FALSE)
}
\arguments{
\item{model_ref}{the model reference}

\item{model_prefix}{the model prefix}

\item{fname_data}{the name of the RData file storing dt_all}

\item{folds_train}{training folds (for rebasing if used)}

\item{whereStatement}{the where statement used during training}

\item{rebase}{whether or not to rebase predictions based on 
difference between mean actual and predictions on train}
}
\value{
nothing is returned.  dt_preds is updated and saved
}
\description{
Create scores / predictions, given a glmnet model and the underlying
mars and categorical variables chosen by IV
}
