% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_glam.R
\name{fn_downsample}
\alias{fn_downsample}
\title{Downsample a data table returning selected columns only}
\usage{
fn_downsample(targetVar, varsToUse, varsToAdd, dt_, method, n_max)
}
\arguments{
\item{targetVar}{the target variable.  The target variable must be 0-1}

\item{varsToUse}{a vector of variables to use
(excluding the missing indicators)}

\item{varsToAdd}{variables to add to varsToUse, typically the
missingInds - the missing indicators}

\item{dt_}{a (training) data table to downsample}

\item{method}{one of caret, imbalanced, reduce.  caret is used to downsample
the majority class (currently assumed to be the negative-0 class)
to the minority class.  Imbalanced downsamples but maintains a
minimum number of the majority class. Reduce simply reduced both
classes in the same proportion leaving n_max observations}

\item{n_max}{the maximum number of examples left by 'reduce', downsampling}
}
\value{
a downsampled dataset
}
\description{
Downsample a data table returning selected columns only
}
